backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.6.0
python-decouple>=3.8
google-generativeai>=0.3.0
tweepy>=4.14.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-environ>=0.11.0
debugpy>=1.8.0